# swaggerui/swagger.yml
openapi: 3.0.0
info:
  description: This is documentation to use Dorayaki backend API
  version: 1.0.0
  title: Swagger Dorayaki backend API
  contact:
    email: 13519105@std.stei.itb.ac.id
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /dorayakis:
    get:
      tags:
        - Dorayakis
      summary: Get all dorayaki's variants.
      description: Get all dorayaki's variants.
      responses:
        200:
          description: Fetching success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DorayakiDataResponse'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
    post:
      tags:
        - Dorayakis
      summary: Create a new Dorayaki variant and save it to database.
      description: Create a new Dorayaki variant using request body. Saving without image.
      security:
        - BearerAuth: []
          CookieAuth: []
          QueryAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputDorayaki'
              description: Input dorayaki variant's data
      responses:
        201:
          description: Create dorayaki success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DorayakiDataResponse'
        400:
          description: Error because invalid body request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        401:
          description: Error because wthere is no JWT token. Check if its exist and valid in order of Bearer, Cookie, and then Query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /dorayakis/search:
    get:
      tags:
        - Dorayakis
        - Paginate
      summary: Paginate dorayaki based on query rasa
      description: Paginate shop based on query sort, itemsPerPage, pageIndex and rasa.
      parameters:
        - in: query
          name: sort
          schema:
            type: string
          description: Determine how data be sorted with name of column. Prefix "-" means descending
        - in: query
          name: itemsPerPage
          schema:
            type: integer
            default: 10
            minimum: 1
          description: Amount of items in 1 page
        - in: query
          name: pageIndex
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Index of current page (1 based index)
        - in: query
          name: kecamatan
          schema:
            type: string
          description: Rasa dorayaki that want to be searched.
      responses:
        200:
          description: Get pagination success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginateDorayakiResponse'
        404:
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotFound'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /dorayakis/{dorayakiId}:
    get:
      tags:
        - Dorayakis
      summary: Get dorayaki by ID
      description: Get dorayaki by ID
      operationId: getDorayaki
      parameters:
        - in: path
          name: dorayakiId
          required: true
          description: ID of the dorayaki
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Get dorayaki success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DorayakiDataResponse'
        404:
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotFound'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
    put:
      tags:
       - Dorayakis
      summary: Edit dorayaki by id
      description: Edit base information about dorayaki
      operationId: updateDorayaki
      security:
        - BearerAuth: []
          CookieAuth: []
          QueryAuth: []
      parameters:
        - in: path
          name: dorayakiId
          required: true
          description: ID of the dorayaki
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputDorayaki'
              description: Input new dorayaki's data
      responses:
        200:
          description: Update dorayaki success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DorayakiDataResponse'
        400:
          description: Error because invalid body request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        401:
          description: Error because wthere is no JWT token. Check if its exist and valid in order of Bearer, Cookie, and then Query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
    delete:
      tags:
       - Dorayakis
      summary: Delete dorayaki by id
      description: Delete dorayaki by id
      operationId: deleteDorayaki
      security:
        - BearerAuth: []
          CookieAuth: []
          QueryAuth: []
      parameters:
        - in: path
          name: dorayakiId
          required: true
          description: ID of the dorayaki
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Delete success
        404:
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotFound'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /dorayakis/{dorayakiId}/upload:
    post:
      tags:
        - Dorayakis
        - Images
      summary: Upload dorayaki image
      description: Upload dorayaki image into filesystem
      operationId: updateDorayaki
      security:
        - BearerAuth: []
          CookieAuth: []
          QueryAuth: []
      parameters:
        - in: path
          name: dorayakiId
          required: true
          description: ID of the dorayaki
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                uploadFile:
                  type: string
                  format: binary
      responses:
        200:
          description: Upload dorayaki success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DorayakiDataResponse'
        400:
          description: Error because invalid body request (exceeding file size)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        401:
          description: Error because wthere is no JWT token. Check if its exist and valid in order of Bearer, Cookie, and then Query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
    QueryAuth:
      type: apiKey
      in: query
      name: jwt
  schemas:
...
    InputDorayaki:
      description: Schema for input dorayaki
      type: object
      properties:
        rasa:
          type: string
          description: Taste of dorayaki
        deskripsi:
          type: string
          description: Description of dorayaki
        image_url:
          type: string
          description: Image URL of dorayaki
      required:
        - rasa
        - deskripsi
    Dorayaki:
      description: Schema for dorayaki
      type: object
      properties:
        id:
          type: integer
          description: ID of dorayaki
        rasa:
          type: string
          description: Taste of dorayaki
        deskripsi:
          type: string
          description: Description of dorayaki
        image_url:
          type: string
          description: Image URL of dorayaki
        created_at:
          type: integer
          description: Time since dorayaki's variant created
        updated_at:
          type: integer
          description: Time since dorayaki's variant updated
      required:
        - rasa
        - deskripsi
...
    ErrorBadRequest:
      description: Bad request
      type: object
      properties:
        status_code:
          type: integer
          description: Error's status code
          default: 400
        message:
          type: string
          description: Error's message
          default: Bad request
    ErrorUnauthorized:
      description: There is no token in header, cookie, or url query
      type: object
      properties:
        status_code:
          type: integer
          description: Error's status code
          default: 401
        message:
          type: string
          description: Error's message
          default: Unauthorized
    ErrorNotFound:
      description: Resource not found
      type: object
      properties:
        status_code:
          type: integer
          description: Error's status code
          default: 404
        message:
          type: string
          description: Error's message
          default: Resource not found
    ErrorInternalServer:
      description: Internal server error
      type: object
      properties:
        status_code:
          type: integer
          description: Error's status code
          default: 500
        message:
          type: string
          description: Error's message
          default: Internal server error
...
    DorayakiDataResponse:
      description: Response that its data contains Dorayaki schema.
      type: object
      properties:
        status_code:
          type: integer
          description: Status code of response
          default: 200
        message:
          type: string
          description: Message of response
          default: Success
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dorayaki'
...
    PaginateDorayakiResponse:
      description: Response for shop/toko pagination
      type: object
      properties:
        status_code:
          type: integer
          description: Status code of response
          default: 200
        message:
          type: string
          description: Message of response
          default: Success
        items_per_page:
          type: integer
          description: Amount of items in 1 page
          default: 10
        total_items:
          type: integer
          description: Total amount of items based on query
        page_index:
          type: integer
          description: Index of current page (1 based index)
          default: 1
        total_pages:
          type: integer
          description: Total pages of pagination result
        sort:
          type: string
          description: Determine how data be sorted with name of column. Prefix "-" means descending
        rasa:
          type: string
          description: Query rasa from input
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dorayaki'
...